+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Fixes `delivery_date` property typo on Rate object
 - Fixes `customs_item.value` to ve a `number` instead of a `string`
 - Add support for `columns` and `additional_columns` on Report creation
+- Adds `address.verifyAddress()` and `Address.createAndVerify()` functions
 
 ## 4.0.0 2021-10-06